routing fix

+++ src/routes/users.js
@@ -1,52 +1,58 @@
 import buildFormObj from '../lib/formObjectBuilder';
-import log from '../lib/logger';
 
-export default (router, { User }) => {
+export default (router, { logger, User }) => {
   router
-    .get('users-all', '/users', async (ctx) => {
-      log('GET /users || All users page');
+    .get('users#index', '/users', async (ctx) => {
+      logger('GET /users || All users page');
       const users = await User.findAll();
-      ctx.render('users/all', { users });
+      ctx.render('users/index', { users });
     })
 
-    .get('users-new', '/users/new', (ctx) => {
-      log('GET /users/new || Sign up page');
+    .get('users#new', '/users/new', (ctx) => {
+      logger('GET /users/new || Sign up page');
       const user = User.build();
       ctx.render('users/new', { f: buildFormObj(user) });
     })
 
-    .get('users-show', '/users/:id', async (ctx) => {
-      log(`GET /users/${ctx.params.id} || Show user`);
+    .get('users#show', '/users/:id', async (ctx) => {
+      logger(`GET /users/${ctx.params.id} || Show user page`);
       const user = await User.findById(ctx.params.id);
-      ctx.render('users/show', { user });
-    })
-
-    .get('users-edit', '/users/:id/edit', async (ctx) => {
-      log(`GET /users/${ctx.params.id}/edit || Edit user`);
-      log(ctx.state.id === ctx.params.id);
-      if (ctx.state.id && Number(ctx.state.id) === Number(ctx.params.id)) {
-        const user = await User.findById(ctx.params.id);
-        log(user);
-        ctx.render('users/edit', { user, f: buildFormObj(user) });
+      if (user) {
+        ctx.render('users/show', { user });
       } else {
-        ctx.redirect(router.url('users-show', ctx.params.id));
+        logger(`GET /users/${ctx.params.id} || Unable to find user`);
+        ctx.throw(404);
       }
     })
 
-    .post('users-create', '/users', async (ctx) => {
-      log('POST /users || User creation');
+    .get('users#edit', '/users/:id/edit', async (ctx) => {
+      logger(`GET /users/${ctx.params.id}/edit || Edit user page`);
+      if (ctx.state.id && Number(ctx.state.id) === Number(ctx.params.id)) {
+        const user = await User.findById(ctx.params.id);
+        ctx.render('users/edit', { user, f: buildFormObj(user) });
+      } else {
+        ctx.flash.set({
+          type: 'danger',
+          text: 'A profile can only be edited by its owner',
+        });
+        ctx.redirect(router.url('users#show', ctx.params.id));
+      }
+    })
+
+    .post('users#create', '/users', async (ctx) => {
+      logger('POST /users || User creation');
       const { form } = ctx.request.body;
       const user = await User.build(form);
       try {
         await user.save();
-        log('POST /users || A user has been created');
+        logger('POST /users || A user has been created');
         ctx.flash.set({
           type: 'success',
           text: `${form.firstName} ${form.lastName} has been created`,
         });
-        ctx.redirect(router.url('users-show', user.dataValues.id));
+        ctx.redirect(router.url('users#show', user.dataValues.id));
       } catch (err) {
-        log('POST /users || Error encountered', err);
+        logger('POST /users || Error encountered', err);
         ctx.state.flash = {
           get: () => ({
             type: 'danger',
@@ -58,14 +64,14 @@ export default (router, { User }) => {
       }
     })
 
-    .patch('users-patch', '/users/:id/edit', async (ctx) => {
-      log('PATCH /users || Editing user info');
+    .patch('users#update', '/users/:id/edit', async (ctx) => {
+      logger('PATCH /users || Editing user info');
       const { form } = ctx.request.body;
       const user = await User.findById(ctx.params.id);
       const { firstName, lastName } = user;
       try {
         await user.update(form, { where: { id: ctx.params.id } });
-        log('PATCH /users || User info has been successfully edited');
+        logger('PATCH /users || User info has been successfully edited');
         ctx.state.flash = {
           get: () => ({
             type: 'success',
@@ -74,7 +80,7 @@ export default (router, { User }) => {
         };
         ctx.render('users/edit', { user, f: buildFormObj(user) });
       } catch (err) {
-        log('PATCH /users || Error encountered', err);
+        logger('PATCH /users || Error encountered', err);
         ctx.state.flash = {
           get: () => ({
             type: 'danger',
@@ -86,20 +92,29 @@ export default (router, { User }) => {
       }
     })
 
-    .delete('users-delete', '/users/:id', async (ctx) => {
-      log('DELETE /users || Delete user');
-      const { firstName, lastName } = await User.findById(ctx.params.id);
-      await User.destroy({
-        where: {
-          id: ctx.params.id,
-        },
-      });
-      log('DELETE /users || User has been deleted');
-      ctx.session = {};
-      ctx.flash.set({
-        type: 'success',
-        text: `${firstName} ${lastName}'s profile has been deleted`,
-      });
-      ctx.redirect(router.url('users-all'));
+    .delete('users#destroy', '/users/:id', async (ctx) => {
+      logger('DELETE /users || Delete user');
+
+      if (ctx.state.id && Number(ctx.state.id) === Number(ctx.params.id)) {
+        const { firstName, lastName } = await User.findById(ctx.params.id);
+        await User.destroy({
+          where: {
+            id: ctx.params.id,
+          },
+        });
+        logger('DELETE /users || User has been deleted');
+        ctx.session = {};
+        ctx.flash.set({
+          type: 'success',
+          text: `${firstName} ${lastName}'s profile has been deleted`,
+        });
+        ctx.redirect(router.url('home'));
+      } else {
+        ctx.flash.set({
+          type: 'danger',
+          text: 'A profile can only be deleted by its owner',
+        });
+        ctx.redirect(router.url('users#show', ctx.params.id));
+      }
     });
 };
